Migrate to Probot 11 and refactor test base
With this change several dependencies including typescript and probot
will be bumped. This requires to make use of octokit instead of github,
which had been changed in the probot module.

The mocking of the API calls had to be adapted when a custom user config
exists, as this had been changed. In addition, the test base has been
refactored and a test helper introduced in order to avoid large code
duplications.